+++ src/index.js
@@ -1,10 +1,10 @@
-import * as sample from "./layouts/sample";
+import * as sample from './layouts/samples';
 import './index.scss';
 
 sample.render();
 
 if (module.hot) {
-  module.hot.accept('./layouts/sample', () => {
+  module.hot.accept('./layouts/samples', () => {
     sample.clean();
     sample.render();
   });
