+++ src/layouts/samples/ImageArea/ImageArea.html
@@ -1,4 +1,4 @@
-<div class='sample-area image-area'>
+<section class='sample-area image-area'>
   <h3>Images by loaders</h3>
 
   <ul class='sample-list'>
@@ -17,4 +17,4 @@
       </div>
     </li>
   </ul>
-</div>
+</section>
