+++ src/layouts/samples/ImageArea/ImageArea.scss
