+++ src/layouts/samples/ImageArea/copyright.html
