+++ src/layouts/samples/ImageArea/index.js
