+++ src/layouts/samples/LazyArea/LazyArea.html
@@ -1,4 +1,4 @@
-<div id='sample-lazy-load'>
+<section id='sample-lazy-load'>
   <h3>Lazy loading</h3>
   <button id='lazy-load-button'>get or load content</button>
-</div>
+</section>
