+++ src/layouts/samples/LazyArea/LazyContent.js
