+++ src/layouts/samples/LazyArea/index.js
