+++ src/layouts/samples/SampleArea.js
